+++ src/input/midiController.ts
@@ -1,10 +1,7 @@
-import {
-  midiNumberToPianoNote,
-} from '../data/piano'
 import { InputLayer } from './inputLayer'
+import { MidiNoteController } from './midiNoteController'
 import {
   parseMidiNoteMessage,
-  type MidiNoteMessage,
 } from '../midi/webMidi'
 
 /**
@@ -12,11 +9,10 @@ import {
  * MIDI-specific state and cleanup stay outside the Input Layer.
  */
 export class MidiInputController {
-  private readonly activeNotes = new Map<number, string>()
-  private readonly inputLayer: InputLayer
+  private readonly midiNoteController: MidiNoteController
 
   constructor(inputLayer: InputLayer) {
-    this.inputLayer = inputLayer
+    this.midiNoteController = new MidiNoteController(inputLayer)
   }
 
   handleMessage = (event: MIDIMessageEvent) => {
@@ -31,41 +27,10 @@ export class MidiInputController {
     }
 
     console.log("[MIDI Monitor]", message)
-    this.handleNoteMessage(message)
+    this.midiNoteController.handleMessage(message)
   }
 
   reset = () => {
-    for (const noteName of this.activeNotes.values()) {
-      this.inputLayer.releaseNote(noteName)
-    }
-
-    this.activeNotes.clear()
-  }
-
-  private handleNoteMessage(message: MidiNoteMessage) {
-    const note = midiNumberToPianoNote(message.noteNumber)
-
-    if (!note) {
-      return
-    }
-
-    if (message.type === "Note On") {
-      if (this.activeNotes.has(message.noteNumber)) {
-        return
-      }
-
-      this.activeNotes.set(message.noteNumber, note.name)
-      this.inputLayer.pressNote(note.name)
-      return
-    }
-
-    const activeNoteName = this.activeNotes.get(message.noteNumber)
-
-    if (!activeNoteName) {
-      return
-    }
-
-    this.activeNotes.delete(message.noteNumber)
-    this.inputLayer.releaseNote(activeNoteName)
+    this.midiNoteController.reset()
   }
 }